Move scipy helpers to their own file.
